+++ tests/Integration/CachedBuilder/PaginateTest.php
@@ -42,7 +42,6 @@ class PaginateTest extends IntegrationTestCase
         $this->assertEquals($liveResults->pluck("name"), $authors->pluck("name"));
     }
 
-    /** @group test */
     public function testPaginationReturnsCorrectLinks()
     {
         if (starts_with(app()->version(), "5.6")